pyproject.toml
[tool.poetry]
name = "recograph"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
black = "^24.10.0"
flake8 = "^7.1.1"
networkx = "^3.4.2"
rest-framework-simplejwt = "^0.0.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
